Handle expand/collapse groups
