+++ packages/module/src/elements/BaseNode.ts
@@ -168,6 +168,23 @@ export default class BaseNode<E extends NodeModel = NodeModel, D = any> extends 
     }, []);
   }
 
+  getPositionableChildren(): Node[] {
+    return super.getChildren().reduce((total, nexChild) => {
+      if (isNode(nexChild)) {
+        if (nexChild.isGroup() && !nexChild.isCollapsed()) {
+          return total.concat(nexChild.getAllNodeChildren());
+        }
+        total.push(nexChild);
+      }
+      return total;
+    }, []);
+  }
+
+  // Return all children regardless of collapse status
+  protected getAllChildren(): GraphElement[] {
+    return super.getChildren();
+  }
+
   getKind(): ModelKind {
     return ModelKind.node;
   }
@@ -202,7 +219,7 @@ export default class BaseNode<E extends NodeModel = NodeModel, D = any> extends 
   updateChildrenPositions(point: Point, prevLocation: Point): void {
     const xOffset = point.x - prevLocation.x;
     const yOffset = point.y - prevLocation.y;
-    this.getAllNodeChildren().forEach(child => {
+    this.getPositionableChildren().forEach(child => {
       if (isNode(child)) {
         const node = child as Node;
         const position = node.getPosition();
