+++ packages/module/src/pipelines/components/groups/DefaultTaskGroup.tsx
@@ -1,209 +1,176 @@
 import * as React from 'react';
 import { observer } from 'mobx-react';
-import { css } from '@patternfly/react-styles';
-import styles from '../../../css/topology-components';
-import CollapseIcon from '@patternfly/react-icons/dist/esm/icons/compress-alt-icon';
-import NodeLabel from '../../../components/nodes/labels/NodeLabel';
-import { Layer } from '../../../components/layers';
-import { GROUPS_LAYER, TOP_LAYER } from '../../../const';
-import { maxPadding, useCombineRefs, useHover } from '../../../utils';
-import { BadgeLocation, GraphElement, isGraph, isNode, LabelPosition, Node, NodeStyle } from '../../../types';
-import {
-  useDragNode,
-  WithContextMenuProps,
-  WithDndDropProps,
-  WithDragNodeProps,
-  WithSelectionProps
-} from '../../../behavior';
-import { CollapsibleGroupProps } from '../../../components';
+import { OnSelect, WithDndDragProps, ConnectDragSource, ConnectDropTarget, WithSelectionProps } from '../../../behavior';
+import { ShapeProps } from '../../../components';
+import { Dimensions } from '../../../geom';
+import { GraphElement, LabelPosition, BadgeLocation, isNode, Node } from '../../../types';
+import { getEdgesFromNodes, getSpacerNodes } from '../../utils';
+import DefaultTaskGroupCollapsed from './DefaultTaskGroupCollapsed';
+import DefaultTaskGroupExpanded from './DefaultTaskGroupExpanded';
 
-type DefaultTaskGroupProps = {
+export interface EdgeCreationTypes {
+  spacerNodeType?: string,
+  edgeType?: string;
+  spacerEdgeType?: string;
+  finallyNodeTypes?: string[];
+  finallyEdgeType?: string;
+}
+
+interface PipelinesDefaultGroupProps {
+  /** Additional content added to the node */
+  children?: React.ReactNode;
+  /** Additional classes added to the group */
   className?: string;
+  /** The graph group node element to represent */
   element: GraphElement;
+  /** Flag if the node accepts drop operations */
   droppable?: boolean;
+  /** Flag if the current drag operation can be dropped on the node */
   canDrop?: boolean;
+  /** Flag if the node is the current drop target */
   dropTarget?: boolean;
+  /** Flag if the user is dragging the node */
   dragging?: boolean;
+  /** Flag if drag operation is a regroup operation */
+  dragRegroupable?: boolean;
+  /** Flag if the user is hovering on the node */
   hover?: boolean;
+  /** Label for the node. Defaults to element.getLabel() */
   label?: string; // Defaults to element.getLabel()
+  /** Secondary label for the node */
   secondaryLabel?: string;
+  /** Flag to show the label */
   showLabel?: boolean; // Defaults to true
+  /** Position of the label, top or bottom. Defaults to element.getLabelPosition() or bottom */
   labelPosition?: LabelPosition;
-  truncateLength?: number; // Defaults to 13
-  badge?: string;
-  badgeColor?: string;
-  badgeTextColor?: string;
-  badgeBorderColor?: string;
-  badgeClassName?: string;
-  badgeLocation?: BadgeLocation;
-  labelOffset?: number; // Space between the label and the group
-  labelIconClass?: string; // Icon to show in label
+  /** The maximum length of the label before truncation */
+  truncateLength?: number;
+  /** The Icon class to show in the label, ignored when labelIcon is specified */
+  labelIconClass?: string;
+  /** The label icon component to show in the label, takes precedence over labelIconClass */
   labelIcon?: string;
+  /** Padding for the label's icon */
   labelIconPadding?: number;
-} & Partial<CollapsibleGroupProps & WithDragNodeProps & WithSelectionProps & WithDndDropProps & WithContextMenuProps>;
+  /** Text for the label's badge */
+  badge?: string;
+  /** Color to use for the label's badge background */
+  badgeColor?: string;
+  /** Color to use for the label's badge text */
+  badgeTextColor?: string;
+  /** Color to use for the label's badge border */
+  badgeBorderColor?: string;
+  /** Additional classes to use for the label's badge */
+  badgeClassName?: string;
+  /** Location of the badge relative to the label's text, inner or below */
+  badgeLocation?: BadgeLocation;
+  /** Flag if the group is collapsible */
+  collapsible?: boolean;
+  /** Width of the collapsed group */
+  collapsedWidth?: number;
+  /** Height of the collapsed group */
+  collapsedHeight?: number;
+  /** Callback when the group is collapsed */
+  onCollapseChange?: (group: Node, collapsed: boolean) => void;
+  /** Shape of the collapsed group */
+  getCollapsedShape?: (node: Node) => React.FunctionComponent<ShapeProps>;
+  /** Shadow offset for the collapsed group */
+  collapsedShadowOffset?: number;
+  /** Flag if the element selected. Part of WithSelectionProps */
+  selected?: boolean;
+  /** Function to call when the element should become selected (or deselected). Part of WithSelectionProps */
+  onSelect?: OnSelect;
+  /** A ref to add to the node for dragging. Part of WithDragNodeProps */
+  dragNodeRef?: WithDndDragProps['dndDragRef'];
+  /** A ref to add to the node for drag and drop. Part of WithDndDragProps */
+  dndDragRef?: ConnectDragSource;
+  /** A ref to add to the node for dropping. Part of WithDndDropProps */
+  dndDropRef?: ConnectDropTarget;
+  /** Function to call to show a context menu for the node  */
+  onContextMenu?: (e: React.MouseEvent) => void;
+  /** Flag indicating that the context menu for the node is currently open  */
+  contextMenuOpen?: boolean;
+  /** Function to return types used to re-create edges on a group collapse/expand (should be the same as calls to getEdgesFromNodes) */
+  getEdgeCreationTypes?: () => {
+    spacerNodeType?: string,
+    edgeType?: string;
+    spacerEdgeType?: string;
+    finallyNodeTypes?: string[];
+    finallyEdgeType?: string;
+  };
+}
 
-type DefaultTaskGroupInnerProps = Omit<DefaultTaskGroupProps, 'element'> & { element: Node };
+type PipelinesDefaultGroupInnerProps = Omit<PipelinesDefaultGroupProps, 'element'> & { element: Node } & WithSelectionProps;
 
-const DefaultTaskGroupInner: React.FunctionComponent<DefaultTaskGroupInnerProps> = observer(
-  ({
-    className,
-    element,
-    collapsible,
-    selected,
-    onSelect,
-    hover,
-    label,
-    secondaryLabel,
-    showLabel = true,
-    truncateLength,
-    canDrop,
-    dropTarget,
-    onContextMenu,
-    contextMenuOpen,
-    dragging,
-    dragNodeRef,
-    badge,
-    badgeColor,
-    badgeTextColor,
-    badgeBorderColor,
-    badgeClassName,
-    badgeLocation,
-    labelOffset = 17,
-    labelIconClass,
-    labelIcon,
-    labelIconPadding,
-    onCollapseChange
-  }) => {
-    const [hovered, hoverRef] = useHover();
-    const [labelHover, labelHoverRef] = useHover();
-    const dragLabelRef = useDragNode()[1];
-    const refs = useCombineRefs<SVGPathElement>(hoverRef, dragNodeRef);
-    const isHover = hover !== undefined ? hover : hovered;
-    const labelPosition = element.getLabelPosition();
+const DefaultTaskGroupInner: React.FunctionComponent<PipelinesDefaultGroupInnerProps> = observer(
+  ({ className, element, onCollapseChange, getEdgeCreationTypes, ...rest }) => {
+    const childCount = element.getAllNodeChildren().length;
 
-    let parent = element.getParent();
-    let altGroup = false;
-    while (!isGraph(parent)) {
-      altGroup = !altGroup;
-      parent = parent.getParent();
-    }
-
-    const children = element.getNodes().filter((c) => c.isVisible());
-
-    // cast to number and coerce
-    const padding = maxPadding(element.getStyle<NodeStyle>().padding ?? 17);
-
-    const { minX, minY, maxX, maxY } = children.reduce(
-      (acc, child) => {
-        const bounds = child.getBounds();
-        return {
-          minX: Math.min(acc.minX, bounds.x - padding),
-          minY: Math.min(acc.minY, bounds.y - padding),
-          maxX: Math.max(acc.maxX, bounds.x + bounds.width + padding),
-          maxY: Math.max(acc.maxY, bounds.y + bounds.height + padding)
-        };
-      },
-      { minX: Infinity, minY: Infinity, maxX: 0, maxY: 0 }
-    );
-
-    const [labelX, labelY] = React.useMemo(() => {
-      if (!showLabel || !(label || element.getLabel())) {
-        return [0, 0];
+    const handleCollapse = (group: Node, collapsed: boolean): void => {
+      if (collapsed && rest.collapsedWidth !== undefined && rest.collapsedHeight !== undefined) {
+        group.setDimensions(new Dimensions(rest.collapsedWidth, rest.collapsedHeight));
       }
-      switch (labelPosition) {
-        case LabelPosition.top:
-          return [minX + (maxX - minX) / 2, -minY + labelOffset];
-        case LabelPosition.right:
-          return [maxX + labelOffset, minY + (maxY - minY) / 2];
-        case LabelPosition.bottom:
-        default:
-          return [minX + (maxX - minX) / 2, maxY + labelOffset];
+      group.setCollapsed(collapsed);
+
+      const controller = group.hasController() && group.getController();
+      if (controller) {
+        const model = controller.toModel();
+        const creationTypes: EdgeCreationTypes = getEdgeCreationTypes ? getEdgeCreationTypes() : {};
+
+        const pipelineNodes = model.nodes.filter((n) => n.type !== creationTypes.spacerNodeType).map((n) => ({ ...n, visible: true }));
+        const spacerNodes = getSpacerNodes(pipelineNodes, creationTypes.spacerNodeType, creationTypes.finallyNodeTypes);
+        const nodes = [...pipelineNodes, ...spacerNodes];
+        const edges = getEdgesFromNodes(
+          pipelineNodes,
+          creationTypes.spacerNodeType,
+          creationTypes.edgeType,
+          creationTypes.edgeType,
+          creationTypes.finallyNodeTypes,
+          creationTypes.finallyEdgeType
+        );
+        controller.fromModel({nodes, edges}, true);
+        controller.getGraph().layout();
       }
-    }, [element, label, labelOffset, labelPosition, maxX, maxY, minX, minY, showLabel]);
+      onCollapseChange && onCollapseChange(group, collapsed);
+    };
 
-    if (children.length === 0) {
-      return null;
+    if (element.isCollapsed()) {
+      return (
+        <DefaultTaskGroupCollapsed
+          className={className}
+          element={element}
+          onCollapseChange={handleCollapse}
+          badge={`${childCount}`}
+          badgeColor="#f5f5f5"
+          badgeBorderColor="#d2d2d2"
+          badgeTextColor="#000000"
+          {...rest}
+        />
+      );
     }
-
-    const groupClassName = css(
-      styles.topologyGroup,
-      className,
-      altGroup && 'pf-m-alt-group',
-      canDrop && 'pf-m-highlight',
-      dragging && 'pf-m-dragging',
-      selected && 'pf-m-selected'
-    );
-    const innerGroupClassName = css(
-      styles.topologyGroup,
-      className,
-      altGroup && 'pf-m-alt-group',
-      canDrop && 'pf-m-highlight',
-      dragging && 'pf-m-dragging',
-      selected && 'pf-m-selected',
-      (isHover || labelHover) && 'pf-m-hover',
-      canDrop && dropTarget && 'pf-m-drop-target'
-    );
-
     return (
-      <g ref={labelHoverRef} onContextMenu={onContextMenu} onClick={onSelect} className={groupClassName}>
-        <Layer id={GROUPS_LAYER}>
-          <g ref={refs} onContextMenu={onContextMenu} onClick={onSelect} className={innerGroupClassName}>
-            <rect
-              x={minX}
-              y={minY}
-              width={maxX - minX}
-              height={maxY - minY}
-              className={styles.topologyGroupBackground}
-            />
-          </g>
-        </Layer>
-        {showLabel && (label || element.getLabel()) && (
-          <Layer id={isHover ? TOP_LAYER : undefined}>
-            <NodeLabel
-              className={styles.topologyGroupLabel}
-              x={labelX}
-              y={labelY}
-              position={labelPosition}
-              paddingX={8}
-              paddingY={5}
-              dragRef={dragNodeRef ? dragLabelRef : undefined}
-              status={element.getNodeStatus()}
-              secondaryLabel={secondaryLabel}
-              truncateLength={truncateLength}
-              badge={badge}
-              badgeColor={badgeColor}
-              badgeTextColor={badgeTextColor}
-              badgeBorderColor={badgeBorderColor}
-              badgeClassName={badgeClassName}
-              badgeLocation={badgeLocation}
-              labelIconClass={labelIconClass}
-              labelIcon={labelIcon}
-              labelIconPadding={labelIconPadding}
-              onContextMenu={onContextMenu}
-              contextMenuOpen={contextMenuOpen}
-              hover={isHover || labelHover}
-              actionIcon={collapsible ? <CollapseIcon /> : undefined}
-              onActionIconClick={() => onCollapseChange(element, true)}
-            >
-              {label || element.getLabel()}
-            </NodeLabel>
-          </Layer>
-        )}
-      </g>
+      <DefaultTaskGroupExpanded
+        className={className}
+        labelPosition={LabelPosition.top}
+        element={element}
+        onCollapseChange={handleCollapse}
+        badgeColor="#f5f5f5"
+        badgeBorderColor="#d2d2d2"
+        badgeTextColor="#000000"
+        {...rest}
+      />
     );
   }
 );
 
-const DefaultTaskGroup: React.FunctionComponent<DefaultTaskGroupProps> = ({
+const DefaultTaskGroup: React.FunctionComponent<PipelinesDefaultGroupProps> = ({
   element,
-  showLabel = true,
-  labelOffset = 17,
   ...rest
-}: DefaultTaskGroupProps) => {
+}: PipelinesDefaultGroupProps) => {
   if (!isNode(element)) {
     throw new Error('DefaultTaskGroup must be used only on Node elements');
   }
-  return <DefaultTaskGroupInner element={element as Node} showLabel={showLabel} labelOffset={labelOffset} {...rest} />;
+
+  return <DefaultTaskGroupInner element={element} {...rest} />;
 };
 
 export default DefaultTaskGroup;
