+++ packages/module/src/pipelines/components/groups/index.ts
@@ -1,4 +1,2 @@
 export { default as DefaultTaskGroup } from './DefaultTaskGroup';
-export { default as PipelinesDefaultGroup } from './PipelinesDefaultGroup';
-export { default as PipelinesDefaultGroupCollapsed } from './PipelinesDefaultGroupCollapsed';
-export { default as PipelinesDefaultGroupExpanded } from './PipelinesDefaultGroupExpanded';
+export type { EdgeCreationTypes } from './DefaultTaskGroup';
